+++ CHANGELOG.md
@@ -49,3 +49,4 @@
 ### 1.1.1
 
 - Updated to NeoForge 21.0.53-beta
+- Added 4 new Genes: Infested, Oozing, Weaving, and Wind Charged